Cleanup of src/backend
- backends take Array<T> as params instead of af_array
- removed Array<T>::eval
- Added createEmptyArray<T>
- Added copyArray<T>
- renamed createView<T> to createSubArray<T>
- Changed moddims to call existing functions
- moved  moddims.*, generator.*, index.* to src/backend/.
- removed the above files from src/backend/*/.

+++ include/af/array.h
@@ -3,7 +3,7 @@
 #include <af/dim4.hpp>
 #define AF_MAX_DIMS 4
 
-typedef long long af_array;
+typedef size_t af_array;
 
 #ifdef __cplusplus
 extern "C" {
